+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added another temporary variabel `tmpIndex` to store the current array index so that the `tmpInts` array can be used without limitations
+- Added support for arrays as function return type
 
 ### Changed